requirements.txt
asgiref==3.8.1
attrs==25.3.0
certifi==2024.8.30
charset-normalizer==3.3.2
Django==4.2
django-filter==24.3
django-ranged-response==0.2.0
django-robots==6.1
django-simple-captcha==0.6.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
idna==3.10
importlib_metadata==8.6.1
importlib_resources==6.5.2
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
Markdown==3.7
packaging==24.2
pillow==11.0.0
PyJWT==2.9.0
pytz==2025.1
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
rpds-py==0.23.1
sqlparse==0.5.1
swagger-spec-validator==3.0.4
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
zipp==3.21.0
